netlink: fix endless loop on 64 bit when parsing binops
mpz_scan1() returns ULONG_MAX when no more bits are found. Due to assignment
to an unsigned int, this value was truncated on 64 bit and the loop never
terminated.

+++ src/netlink_delinearize.c
@@ -555,7 +555,7 @@ static void expr_postprocess(struct rule_pp_ctx *ctx,
 			     struct stmt *stmt, struct expr **exprp)
 {
 	struct expr *expr = *exprp, *i;
-	unsigned int n;
+	unsigned long n;
 
 	//pr_debug("%s len %u\n", expr->ops->name, expr->len);
